server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "msal>=1.34.0",
    "numpy>=2.4.1",
    "loguru>=0.7.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.40.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "qdrant-client>=1.14.0",
    "tuspyserver>=4.2.3",
    "orjson>=3.11.7",
]
